Fix standalone build (#143)
* Revert "Fix"

This reverts commit 40da19c64763fade9733aedaf770077636e5890b.

* Fixes for standalone build

* Fixes

+++ inference-engine/tests/functional/shared_test_classes/include/shared_test_classes/single_layer/variadic_split.hpp
@@ -15,7 +15,7 @@
 namespace LayerTestsDefinitions {
 
 typedef std::tuple<
-        std::vector<int32_t>,           // Num splits
+        std::vector<size_t>,            // Num splits
         size_t,                         // Axis
         InferenceEngine::Precision,     // Net precision
         InferenceEngine::Precision,     // Input precision
